added bull, changed to an hour
