+++ .env.default
@@ -1,7 +1,9 @@
-EMOJI='🌍'
-NODE_ENV='development'
-DEV_LOG_LEVEL='debug'
-PROD_LOG_LEVEL='error'
-
 DISCORD_TOKEN=
 GUILD_ID=
+
+DEV_LOG_LEVEL='debug'
+PROD_LOG_LEVEL='error'
+REDIS_PASSWORD=
+REDIS_USERNAME=
+REDIS_HOST=
+REDIS_PORT=
