+++ package.json
@@ -10,7 +10,7 @@
     "start": "node dist/index.js",
     "serve-dev": "nodemon dist/index.js",
     "serve-debug": "nodemon --inspect dist/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node src/test.js",
     "watch-ts": "tsc -w"
   },
   "keywords": [],
@@ -18,6 +18,7 @@
   "license": "ISC",
   "dependencies": {
     "big.js": "^6.2.1",
+    "bull": "^4.11.3",
     "discord.js": "^14.3.0",
     "dotenv": "^16.0.2",
     "node-cron": "^3.0.2",
@@ -27,6 +28,7 @@
   },
   "devDependencies": {
     "@types/big.js": "^6.1.5",
+    "@types/bull": "^4.10.0",
     "@types/node": "^18.7.14",
     "@types/node-cron": "^3.0.3",
     "@types/pino": "^7.0.4",
